size changed from uint to size_t

+++ Project/Milestone 1/Station.h
@@ -16,7 +16,7 @@ namespace seneca {
 		uint m_serialNumber;
 		uint m_stockQunatity;
 
-		static uint m_widthField;
+		static size_t m_widthField;
 		static uint id_generator;
 
 	public:
